Multi-user alfa-version Tic-tac-toe game is completed.
You can specify the names of the players.
If the names are the same or empty, then the names are chosen by default

Button new game clears scores and it allows you to specify new names players.

At the end of the round are invited to play or start a new